chore: refactor authentication failure redirect and retry (#34899)
## Description
> Add infrastructure for Authentication failure to retry and redirect.
> An additional change has been made to make the Sentry's doLog() method
public for others to use it for logging additional errors to Sentry.


Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.Sanity,@tag.Authentication"

### :mag: Cypress test results
