+++ app/server/appsmith-server/src/main/java/com/appsmith/server/authentication/handlers/AuthenticationFailureHandler.java
@@ -1,9 +1,13 @@
 package com.appsmith.server.authentication.handlers;
 
 import com.appsmith.server.authentication.handlers.ce.AuthenticationFailureHandlerCE;
-import lombok.RequiredArgsConstructor;
+import com.appsmith.server.authentication.helpers.AuthenticationFailureRetryHandler;
 import org.springframework.stereotype.Component;
 
 @Component
-@RequiredArgsConstructor
-public class AuthenticationFailureHandler extends AuthenticationFailureHandlerCE {}
+public class AuthenticationFailureHandler extends AuthenticationFailureHandlerCE {
+
+    public AuthenticationFailureHandler(AuthenticationFailureRetryHandler authenticationFailureRetryHandler) {
+        super(authenticationFailureRetryHandler);
+    }
+}
